+++ R/choices_labeled.R
@@ -131,7 +131,8 @@ choices_labeled <- function(choices, labels, subset = NULL, types = NULL) {
 #'
 #' @param data (`data.frame` or `character`)
 #' If `data.frame`, then data to extract labels from.
-#' If `character`, then name of the dataset to extract data from once available.
+#' If `character`, then name of the dataset to extract data from once available. Keyword `"all"`
+#' (default) indicates that `data` will be inherited from the `data_extract_spec` `dataname`.
 #' @param subset (`character` or `function`)
 #' If `character`, then a vector of column names.
 #' If `function`, then this function is used to determine the possible columns (e.g. all factor columns).
@@ -161,7 +162,10 @@ choices_labeled <- function(choices, labels, subset = NULL, types = NULL) {
 #'   key = default_cdisc_join_keys["ADRS", "ADRS"]
 #' )
 #'
-#' # delayed version
+#' # delayed version with unknown dataset
+#' variable_choices()
+#'
+#' # delayed ADRS
 #' variable_choices("ADRS", subset = c("USUBJID", "STUDYID"))
 #'
 #' # functional subset (with delayed data) - return only factor variables
@@ -171,35 +175,48 @@ choices_labeled <- function(choices, labels, subset = NULL, types = NULL) {
 #' })
 #' @export
 #'
-variable_choices <- function(data, subset = NULL, fill = FALSE, key = NULL) {
+variable_choices <- function(data = "all", subset = function(data) names(data), fill = FALSE, key = NULL) {
   checkmate::assert(
     checkmate::check_character(subset, null.ok = TRUE, any.missing = FALSE),
-    checkmate::check_function(subset)
+    checkmate::check_function(subset, args = "data")
   )
   checkmate::assert_flag(fill)
   checkmate::assert_character(key, null.ok = TRUE, any.missing = FALSE)
-
-  UseMethod("variable_choices")
+  if (is.character(data)) {
+    variable_choices_delayed(data = data, subset = subset, fill = fill, key = key)
+  } else {
+    variable_choices_data_frame(data = data, subset = subset, fill = fill, key = key)
+  }
 }
 
-#' @rdname variable_choices
-#' @export
-variable_choices.character <- function(data, subset = NULL, fill = FALSE, key = NULL) {
+#' @keywords internal
+variable_choices_delayed <- function(data, subset = function(data) names(data), fill = FALSE, key = NULL) {
+  checkmate::assert_string(data)
   structure(list(data = data, subset = subset, key = key),
     class = c("delayed_variable_choices", "delayed_data", "choices_labeled")
   )
 }
 
-#' @rdname variable_choices
-#' @export
-variable_choices.data.frame <- function(data, subset = NULL, fill = TRUE, key = NULL) {
+#' @keywords internal
+variable_choices_data_frame <- function(data, subset = function(data) names(data), fill = TRUE, key = NULL) {
+  checkmate::assert_data_frame(data, min.cols = 1)
   checkmate::assert(
     checkmate::check_character(subset, null.ok = TRUE),
     checkmate::check_function(subset, null.ok = TRUE)
   )
 
   if (is.function(subset)) {
-    subset <- resolve_delayed_expr(subset, ds = data, is_value_choices = FALSE)
+    subset <- subset(data)
+    if (
+      !checkmate::test_character(subset, any.missing = FALSE) ||
+        length(subset) > ncol(data) ||
+        anyDuplicated(subset)
+    ) {
+      stop(
+        "variable_choices(subset) function in must return a character vector with unique",
+        "names from the available columns of the dataset"
+      )
+    }
   }
 
   checkmate::assert_subset(subset, c("", names(data)), empty.ok = TRUE)
@@ -283,8 +300,8 @@ variable_choices.data.frame <- function(data, subset = NULL, fill = TRUE, key = 
 #' })
 #' @export
 #'
-value_choices <- function(data,
-                          var_choices,
+value_choices <- function(data = "all",
+                          var_choices = variable_choices(data = data),
                           var_label = NULL,
                           subset = NULL,
                           sep = " - ") {
@@ -298,16 +315,21 @@ value_choices <- function(data,
     checkmate::check_function(subset)
   )
   checkmate::assert_string(sep)
-  UseMethod("value_choices")
+
+  if (is.character(data)) {
+    value_choices_delayed(data = data, var_choices = var_choices, var_label = var_label, subset = subset, sep = sep)
+  } else {
+    value_choices_data_frame(data = data, var_choices = var_choices, var_label = var_label, subset = subset, sep = sep)
+  }
 }
 
-#' @rdname value_choices
-#' @export
-value_choices.character <- function(data,
-                                    var_choices,
-                                    var_label = NULL,
-                                    subset = NULL,
-                                    sep = " - ") {
+#' @keywords internal
+value_choices_delayed <- function(data,
+                                  var_choices,
+                                  var_label = NULL,
+                                  subset = NULL,
+                                  sep = " - ") {
+  checkmate::assert_string(data, null.ok = TRUE)
   structure(
     list(
       data = data,
@@ -320,13 +342,13 @@ value_choices.character <- function(data,
   )
 }
 
-#' @rdname value_choices
-#' @export
-value_choices.data.frame <- function(data,
+#' @keywords internal
+value_choices_data_frame <- function(data,
                                      var_choices,
                                      var_label = NULL,
                                      subset = NULL,
                                      sep = " - ") {
+  checkmate::assert_data_frame(data, min.cols = 1)
   checkmate::assert_subset(var_choices, names(data))
   checkmate::assert_subset(var_label, names(data), empty.ok = TRUE)
 
@@ -369,7 +391,13 @@ value_choices.data.frame <- function(data,
   df <- unique(data.frame(choices, labels, stringsAsFactors = FALSE)) # unique combo of choices x labels
 
   if (is.function(subset)) {
-    subset <- resolve_delayed_expr(subset, ds = data, is_value_choices = TRUE)
+    subset <- subset(data)
+    if (!checkmate::test_atomic(subset) || anyDuplicated(subset)) {
+      stop(
+        "value_choices(subset) function must return a vector with unique values from the ",
+        "respective columns of the dataset."
+      )
+    }
   }
   res <- choices_labeled(
     choices = df$choices,