+++ R/data_extract_spec.R
@@ -15,7 +15,8 @@
 #' `teal.transform` uses this object to construct a UI element in a module.
 #'
 #' @param dataname (`character`)
-#' The name of the dataset to be extracted.
+#' The name of the dataset to be extracted. Keyword `"all"` suggest that this `data_extract_spec`
+#' is universal for all datasets used in `teal` application.
 #' @param select (`NULL` or `select_spec`-S3 class or `delayed_select_spec`)
 #' Columns to be selected from the input dataset mentioned in `dataname`.
 #' The setup can be created using [select_spec] function.
@@ -83,7 +84,10 @@
 #' )
 #' @export
 #'
-data_extract_spec <- function(dataname, select = NULL, filter = NULL, reshape = FALSE) {
+data_extract_spec <- function(dataname = "all",
+                              select = select_spec(selected = all_choices()), # todo: remove this default
+                              filter = filter_spec(), # todo: default filter should be NULL
+                              reshape = FALSE) {
   checkmate::assert_string(dataname)
   stopifnot(
     is.null(select) ||
@@ -96,22 +100,11 @@ data_extract_spec <- function(dataname, select = NULL, filter = NULL, reshape = 
   )
   checkmate::assert_flag(reshape)
 
-  if (is.null(select) && is.null(filter)) {
-    select <- select_spec(
-      choices = variable_choices(dataname),
-      multiple = TRUE
-    )
-    filter <- filter_spec(
-      vars = choices_selected(variable_choices(dataname)),
-      selected = all_choices()
-    )
-  }
-
   if (inherits(filter, "filter_spec")) filter <- list(filter)
 
-  for (idx in seq_along(filter)) filter[[idx]]$dataname <- dataname
+  for (idx in seq_along(filter)) filter[[idx]]$dataname <- dataname # todo: find where it is used
 
-  if (
+  ans <- if (
     inherits(select, "delayed_select_spec") ||
       any(vapply(filter, inherits, logical(1), "delayed_filter_spec"))
   ) {
@@ -125,4 +118,5 @@ data_extract_spec <- function(dataname, select = NULL, filter = NULL, reshape = 
       class = "data_extract_spec"
     )
   }
+  ans
 }