+++ R/resolve.R
@@ -7,60 +7,49 @@
 #'
 #' @param x (`delayed_data`) object to resolve.
 #' @param datasets (named `list` of `data.frame`) to use in evaluation.
-#' @param keys (named `list` of `character`) to be used as the keys for each dataset.
-#' The names of this list must be exactly the same as for datasets.
+#' @param join_keys (`join_keys`) used to resolve `key` in [variable_choices()].
 #'
 #' @return Resolved object.
 #'
 #' @keywords internal
 #'
-resolve <- function(x, datasets, keys = NULL) {
-  checkmate::assert_list(datasets, types = "reactive", min.len = 1, names = "named")
-  checkmate::assert_list(keys, "character", names = "named", null.ok = TRUE)
-  checkmate::assert(
-    .var.name = "keys",
-    checkmate::check_names(names(keys), subset.of = names(datasets)),
-    checkmate::check_null(keys)
-  )
-
+resolve <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  checkmate::assert_list(datasets, min.len = 1, names = "named")
+  checkmate::assert_class(join_keys, "join_keys")
   UseMethod("resolve")
 }
 
 #' @describeIn resolve Call [variable_choices()] on the delayed `variable_choices` object.
 #' @export
-resolve.delayed_variable_choices <- function(x, datasets, keys) {
+resolve.delayed_variable_choices <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  checkmate::assert_list(datasets, len = 1, names = "named")
+  x$data <- datasets[[1]]
   if (is.null(x$key)) {
-    x$key <- `if`(is.null(keys), character(), keys[[x$data]])
+    dataname <- names(datasets)
+    x$key <- as.character(join_keys[dataname, dataname])
   }
-  x$data <- datasets[[x$data]]()
-  if (inherits(x$subset, "function")) {
-    x$subset <- resolve_delayed_expr(x$subset, ds = x$data, is_value_choices = FALSE)
-  }
-
   do.call("variable_choices", x)
 }
 
 #' @describeIn resolve Call [value_choices()] on the delayed `value_choices` object.
 #' @export
-resolve.delayed_value_choices <- function(x, datasets, keys) {
-  x$data <- datasets[[x$data]]()
+resolve.delayed_value_choices <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  checkmate::assert_list(datasets, len = 1, names = "named")
+  x$data <- datasets[[1]]
   if (inherits(x$var_choices, "delayed_variable_choices")) {
-    x$var_choices <- resolve(x$var_choices, datasets, keys)
+    x$var_choices <- resolve(x$var_choices, datasets, join_keys = join_keys)
   }
-  if (is.function(x$subset)) {
-    x$subset <- resolve_delayed_expr(x$subset, ds = x$data, is_value_choices = TRUE)
-  }
-
   do.call("value_choices", x)
 }
 
 #' @describeIn resolve Call [select_spec()] on the delayed `choices_selected` object.
 #' @export
-resolve.delayed_choices_selected <- function(x, datasets, keys) {
+resolve.delayed_choices_selected <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  checkmate::assert_list(datasets, len = 1, names = "named")
   if (inherits(x$selected, "delayed_data")) {
-    x$selected <- resolve(x$selected, datasets = datasets, keys)
+    x$selected <- resolve(x$selected, datasets = datasets, join_keys = join_keys)
   }
-  x$choices <- resolve(x$choices, datasets = datasets, keys)
+  x$choices <- resolve(x$choices, datasets = datasets, join_keys = join_keys)
 
   if (!all(x$selected %in% x$choices)) {
     warning(paste(
@@ -76,29 +65,30 @@ resolve.delayed_choices_selected <- function(x, datasets, keys) {
 
 #' @describeIn resolve Call [select_spec()] on the delayed specification.
 #' @export
-resolve.delayed_select_spec <- function(x, datasets, keys) {
-  x$choices <- resolve(x$choices, datasets = datasets, keys)
+resolve.delayed_select_spec <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  checkmate::assert_list(datasets, len = 1, names = "named")
+  x$choices <- resolve(x$choices, datasets = datasets, join_keys = join_keys)
   if (inherits(x$selected, "delayed_data")) {
-    x$selected <- resolve(x$selected, datasets = datasets, keys)
+    x$selected <- resolve(x$selected, datasets = datasets, join_keys = join_keys)
   }
-
   do.call("select_spec", x)
 }
 
 #' @describeIn resolve Call [filter_spec()] on the delayed specification.
 #' @export
-resolve.delayed_filter_spec <- function(x, datasets, keys) {
+resolve.delayed_filter_spec <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  checkmate::assert_list(datasets, len = 1, names = "named")
   if (inherits(x$vars_choices, "delayed_data")) {
-    x$vars_choices <- resolve(x$vars_choices, datasets = datasets, keys)
+    x$vars_choices <- resolve(x$vars_choices, datasets = datasets, join_keys = join_keys)
   }
   if (inherits(x$vars_selected, "delayed_data")) {
-    x$vars_selected <- resolve(x$vars_selected, datasets = datasets, keys)
+    x$vars_selected <- resolve(x$vars_selected, datasets = datasets, join_keys = join_keys)
   }
   if (inherits(x$choices, "delayed_data")) {
-    x$choices <- resolve(x$choices, datasets = datasets, keys)
+    x$choices <- resolve(x$choices, datasets = datasets, join_keys = join_keys)
   }
   if (inherits(x$selected, "delayed_data")) {
-    x$selected <- resolve(x$selected, datasets = datasets, keys)
+    x$selected <- resolve(x$selected, datasets = datasets, join_keys = join_keys)
   }
 
   do.call("filter_spec_internal", x[intersect(names(x), methods::formalArgs(filter_spec_internal))])
@@ -106,16 +96,17 @@ resolve.delayed_filter_spec <- function(x, datasets, keys) {
 
 #' @describeIn resolve Call [data_extract_spec()] on the delayed specification.
 #' @export
-resolve.delayed_data_extract_spec <- function(x, datasets, keys) {
+resolve.delayed_data_extract_spec <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  dataset <- datasets[x$dataname]
   x$select <- `if`(
     inherits(x$select, "delayed_data"),
-    resolve(x$select, datasets = datasets, keys),
+    resolve(x$select, datasets = dataset, join_keys = join_keys),
     x$select
   )
 
   if (any(vapply(x$filter, inherits, logical(1L), "delayed_data"))) {
     idx <- vapply(x$filter, inherits, logical(1), "delayed_data")
-    x$filter[idx] <- lapply(x$filter[idx], resolve, datasets = datasets, keys = keys)
+    x$filter[idx] <- lapply(x$filter[idx], resolve, datasets = dataset, join_keys = join_keys)
   }
 
   do.call("data_extract_spec", x)
@@ -124,55 +115,24 @@ resolve.delayed_data_extract_spec <- function(x, datasets, keys) {
 #' @describeIn resolve Iterates over elements of the list and recursively calls
 #' `resolve`.
 #' @export
-resolve.list <- function(x, datasets, keys) {
+resolve.list <- function(x, datasets, join_keys = teal.data::join_keys()) {
+  new_x <- if (checkmate::test_list(x, "data_extract_spec", min.len = 1) && identical(x[[1]]$dataname, "all")) {
+    lapply(seq_along(datasets), function(i) {
+      xx <- x[[1]]
+      xx$dataname <- names(datasets)[i]
+      xx
+    })
+  } else {
+    x
+  }
   # If specified explicitly, return it unchanged. Otherwise if delayed, resolve.
-  lapply(x, resolve, datasets = datasets, keys = keys)
+  lapply(new_x, resolve, datasets = datasets, join_keys = join_keys)
 }
 
 #' @describeIn resolve Default method that does nothing and returns `x` itself.
 #' @export
-resolve.default <- function(x, datasets, keys) {
+resolve.default <- function(x, datasets, join_keys = teal.data::join_keys()) {
   x
-}
-
-#' Resolve expression after delayed data are loaded
-#'
-#'
-#' @param x (`function`) Function that is applied on dataset.
-#' It must take only a single argument "data" and return character vector with columns / values.
-#' @param ds (`data.frame`) Dataset.
-#' @param is_value_choices (`logical`) Determines which check of the returned value will be applied.
-#'
-#' @return `character` vector - result of calling function `x` on dataset `ds`.
-#'
-#' @keywords internal
-#'
-resolve_delayed_expr <- function(x, ds, is_value_choices) {
-  checkmate::assert_function(x, args = "data", nargs = 1)
-
-  # evaluate function
-  res <- do.call(x, list(data = ds))
-
-  # check returned value
-  if (is_value_choices) {
-    if (!checkmate::test_atomic(res) || anyDuplicated(res)) {
-      stop(paste(
-        "The following function must return a vector with unique values",
-        "from the respective columns of the dataset.\n\n",
-        deparse1(bquote(.(x)), collapse = "\n")
-      ))
-    }
-  } else {
-    if (!checkmate::test_character(res, any.missing = FALSE) || length(res) > ncol(ds) || anyDuplicated(res)) {
-      stop(paste(
-        "The following function must return a character vector with unique",
-        "names from the available columns of the dataset:\n\n",
-        deparse1(bquote(.(x)), collapse = "\n")
-      ))
-    }
-  }
-
-  res
 }
 
 #' @export