+++ tests/testthat/test-variable_choices.R
@@ -1,129 +1,58 @@
-ADSL <- teal.data::rADSL
-ADTTE <- teal.data::rADTTE
-data_list <- list(ADSL = reactive(ADSL), ADTTE = reactive(ADTTE))
-primary_keys_list <- list(ADSL = c("STUDYID", "USUBJID"), ADTTE = c("STUDYID", "USUBJID", "PARAMCD"))
-
-test_that("Can create variable_choices with datasets with no or missing labels", {
-  example_data <- data.frame(USUBJID = 1:2, STUDYID = 1:1)
-
-  # no labels given
-  choice_1 <- variable_choices(example_data, fill = TRUE)
-  testthat::expect_equal(names(choice_1), c("USUBJID: USUBJID", "STUDYID: STUDYID"))
-
-  # one missing label
-  missing_one_label_data <- example_data
-  teal.data::col_labels(missing_one_label_data) <- c(as.character(NA), "Label")
-  choice_2 <- variable_choices(missing_one_label_data, fill = FALSE)
-  testthat::expect_equal(names(choice_2), c("USUBJID: Label Missing", "STUDYID: Label"))
-
-  # all missing label
-  missing_two_label_data <- example_data
-  teal.data::col_labels(missing_two_label_data) <- c(as.character(NA), as.character(NA))
-  choice_2 <- variable_choices(missing_two_label_data, fill = FALSE)
-  testthat::expect_equal(names(choice_2), c("USUBJID: Label Missing", "STUDYID: Label Missing"))
+testthat::test_that("variable_choices returns variable_choices", {
+  testthat::expect_s3_class(variable_choices(), "choices_labeled")
 })
 
-test_that("delayed version of variable_choices", {
-  # hard-coded subset
-  obj <- variable_choices("ADSL", subset = c("SEX", "ARMCD", "COUNTRY"))
-  testthat::expect_equal(
-    obj,
-    structure(
-      list(data = "ADSL", subset = c("SEX", "ARMCD", "COUNTRY"), key = NULL),
-      class = c("delayed_variable_choices", "delayed_data", "choices_labeled")
-    )
+testthat::test_that("variable_choices accepts data to be  string or data.frame with cols", {
+  testthat::expect_no_error(variable_choices(data = "all"))
+  testthat::expect_no_error(variable_choices(data = mtcars))
+  testthat::expect_error(variable_choices(data = NULL))
+  testthat::expect_error(variable_choices(data = data.frame()))
+  testthat::expect_error(variable_choices(data = c("a", "b")))
+})
+
+testthat::test_that("variable_choices accepts subset to be character or a function. Both checked against colnames", {
+  testthat::expect_no_error(variable_choices(iris, c("Species", "Sepal.Length")))
+  testthat::expect_no_error(variable_choices(iris, subset = function(data) colnames(data)))
+  testthat::expect_error(variable_choices(iris, c("Species", "idontexist")), "Must be a subset of")
+})
+
+testthat::test_that("variable_choices accepts subset to be character or a function when data is NULL", {
+  testthat::expect_no_error(variable_choices(subset = c("Species", "Sepal.Length")))
+  testthat::expect_no_error(variable_choices(subset = function(data) colnames(data)))
+  testthat::expect_no_error(variable_choices(subset = c("Species", "idontexist")))
+  testthat::expect_no_error(variable_choices(subset = function(data) c("Species", "idontexist")))
+})
+
+testthat::test_that("variable_choices has defaults for all arguments", {
+  testthat::expect_no_error(variable_choices())
+})
+
+testthat::test_that("variable_choices subset function is evaluated when data is eager", {
+  testthat::expect_identical(
+    variable_choices(iris, subset = function(data) names(data)),
+    variable_choices(iris, subset = colnames(iris))
   )
+  testthat::expect_error(variable_choices(iris, subset = function(data) c("idontexist")), "Must be a subset of")
+})
 
-  data_list <- list(ADSL = reactive(ADSL), ADTTE = reactive(ADTTE))
-  key_list <- list(ADSL = c("STUDYID", "USUBJID"), ADTTE = c("STUDYID", "USUBJID", "PARAMCD"))
-
-  res_obj <- isolate(resolve(obj, datasets = data_list, keys = key_list))
-  testthat::expect_equal(
-    res_obj,
-    variable_choices(ADSL, subset = c("SEX", "ARMCD", "COUNTRY"))
+testthat::test_that("resolved delayed_variable_choices is identical to variable_choices specified with data", {
+  testthat::expect_identical(
+    resolve(variable_choices(), datasets = list(iris = iris)),
+    variable_choices(iris, subset = function(data) names(data))
   )
-
-  # functional subset
-  obj <- variable_choices("ADSL", subset = function(data) colnames(data)[1:2])
-  testthat::expect_equal(
-    obj,
-    structure(
-      list(data = "ADSL", subset = function(data) colnames(data)[1:2], key = NULL),
-      class = c("delayed_variable_choices", "delayed_data", "choices_labeled")
-    )
-  )
-
-  res_obj <- isolate(resolve(obj, datasets = data_list, keys = key_list))
-  testthat::expect_equal(
-    res_obj,
-    variable_choices(ADSL, subset = colnames(ADSL)[1:2], key = c("STUDYID", "USUBJID"))
-  )
-
-  # non-null key value
-  obj <- variable_choices("ADSL", key = c("USUBJID", "STUDYID"))
-  testthat::expect_equal(
-    obj,
-    structure(
-      list(data = "ADSL", subset = NULL, key = c("USUBJID", "STUDYID")),
-      class = c("delayed_variable_choices", "delayed_data", "choices_labeled")
-    )
-  )
-
-  res_obj <- isolate(resolve(obj, datasets = data_list, keys = key_list))
-  testthat::expect_equal(
-    res_obj,
-    variable_choices(ADSL, key = c("USUBJID", "STUDYID"))
+  testthat::expect_identical(
+    resolve(variable_choices(subset = c("Species", "Sepal.Length")), datasets = list(iris = iris)),
+    variable_choices(iris, subset = c("Species", "Sepal.Length"))
   )
 })
 
-
-test_that("delayed version of variable_choices - resolve_delayed", {
-  # hard-coded subset
-  obj <- variable_choices("ADSL", subset = c("SEX", "ARMCD", "COUNTRY"))
-  testthat::expect_equal(
-    obj,
-    structure(
-      list(data = "ADSL", subset = c("SEX", "ARMCD", "COUNTRY"), key = NULL),
-      class = c("delayed_variable_choices", "delayed_data", "choices_labeled")
-    )
+testthat::test_that("resolving delayed_variable_choices throws the error when subset isn't a colname", {
+  testthat::expect_error(
+    resolve(variable_choices(subset = function(data) "idontexist"), datasets = list(iris = iris)),
+    "Must be a subset of"
   )
-
-  res_obj <- isolate(resolve_delayed(obj, datasets = data_list, keys = primary_keys_list))
-  testthat::expect_equal(
-    res_obj,
-    variable_choices(ADSL, subset = c("SEX", "ARMCD", "COUNTRY"))
-  )
-
-
-  # functional subset
-  obj <- variable_choices("ADSL", subset = function(data) colnames(data)[1:2])
-  testthat::expect_equal(
-    obj,
-    structure(
-      list(data = "ADSL", subset = function(data) colnames(data)[1:2], key = NULL),
-      class = c("delayed_variable_choices", "delayed_data", "choices_labeled")
-    )
-  )
-
-  res_obj <- isolate(resolve_delayed(obj, datasets = data_list, keys = primary_keys_list))
-  testthat::expect_equal(
-    res_obj,
-    variable_choices(ADSL, subset = colnames(ADSL)[1:2], key = c("STUDYID", "USUBJID"))
-  )
-
-  # non-null key value
-  obj <- variable_choices("ADSL", key = c("USUBJID", "STUDYID"))
-  testthat::expect_equal(
-    obj,
-    structure(
-      list(data = "ADSL", subset = NULL, key = c("USUBJID", "STUDYID")),
-      class = c("delayed_variable_choices", "delayed_data", "choices_labeled")
-    )
-  )
-
-  res_obj <- isolate(resolve_delayed(obj, datasets = data_list, keys = primary_keys_list))
-  testthat::expect_equal(
-    res_obj,
-    variable_choices(ADSL, key = c("USUBJID", "STUDYID"))
+  testthat::expect_error(
+    resolve(variable_choices(subset = "idontexist"), datasets = list(iris = iris)),
+    "Must be a subset of"
   )
 })